requirements.txt
flask
Pillow
flask_sqlalchemy
